Fix incorrect formatting of NSAP area addresses/address prefixes.
If the area address/address prefix is 13 octets long, use the 13th
octet, not the 21st octet; the latter is *not* part of the area
address/address prefix, and might either not be in the packet or might
be some random other part of the packet.

Add/expand comments while we're at it.

Bug: 14744
Change-Id: I7b90318a72a49b67d8ec17952add528185fd064b
Reviewed-on: https://code.wireshark.org/review/27848
Reviewed-by: Guy Harris <guy@alum.mit.edu>
(cherry picked from commit 101b692569f17d14f7165446a8a2ae26415057cb)
Reviewed-on: https://code.wireshark.org/review/27849