Fix test for early exit from MMSE dissector.
We don't need to dissect the body if 1) we're *not* building a protocol
tree (tree == NULL) *and* 2) the PDU doesn't have content that can be
handed off to subdissectors.

(Fix which vs. that issue in a comment while we're at it.)

Change-Id: I90890975c05e72cc9ebc776a21683905828f57b5
Reviewed-on: https://code.wireshark.org/review/27876
Reviewed-by: Guy Harris <guy@alum.mit.edu>
(cherry picked from commit 91c069d5b3dd4736653287a9e454b06354564011)
Reviewed-on: https://code.wireshark.org/review/27877