Move make-taps and make-dissectors to tools/
make-taps and make dissectors are build tools so that is the natural
location for them. This also makes it easier to fix bug 14622 for
autotools. When cross compiling we need to use the host target to
build these executables.

Adjust CMake and autotools builds to reflect the change.

Ping-Bug: 14622
Change-Id: I3034f544dad53fcc83230f5050512260fc1ab198
Reviewed-on: https://code.wireshark.org/review/27920
Petri-Dish: João Valverde <j@v6e.pt>
Reviewed-by: João Valverde <j@v6e.pt>