More debugging for bug 14738.
It *should* be throwing an exception when we add a string with a huge
length, and it *does* throw an exception on my machine, but it doesn't
appear to be doing so on the buildbot.  Try to see why that might be.

Change-Id: Icaf4fc08421d3ceae9aca777282082d198e185cf
Reviewed-on: https://code.wireshark.org/review/27999
Reviewed-by: Guy Harris <guy@alum.mit.edu>
(cherry picked from commit 94390dcb3a4b62a0ba595aa3e92edb31b83de4c2)
Reviewed-on: https://code.wireshark.org/review/28000