Don't use FT_STRING fields for non-strings.
If you're reporting an error, use an expert info item.

If you're putting a structure into the protocol tree, use FT_NONE for
the structure as a whole.

(In 2.2, we're not doing all of them; in later releases, a significant
restructuring was done, which eliminated many of them.)
Change-Id: Ie89b552576b15195acb0a9108d33430115d99f00
Reviewed-on: https://code.wireshark.org/review/28024
Reviewed-by: Guy Harris <guy@alum.mit.edu>
(cherry picked from commit 8e7396edb4b4ac9580cdddb4feb0903da071905f)
Reviewed-on: https://code.wireshark.org/review/28027