Make sure proto_tree_add_expert items are always added to the tree.
Make sure that proto_tree_add_text_internal() and
proto_tree_add_text_valist_internal() don't throw an exception, so the
indication always appears in the tree to indicate the issue.  Do the
"do the bytes exist" check *after* we've added all of the expert info to
the protocol tree, so we still throw the appropriate exception.

Change-Id: I4e0d2dcc48f9c8f4482550ae16284b9e021232cd
Reviewed-on: https://code.wireshark.org/review/28062
Reviewed-by: Guy Harris <guy@alum.mit.edu>