Qt: Set the ByteViewText field and protocol foreground.
Set a foreground color when we add a field and protocol format range. On
some systems the highlight background is darker or lighter than the
normal background, and we need to adjust the foreground to match.

Add a note about correctly handling our hover rect color.

Change-Id: Ib6b50403dec889280eb3d27f3deb10150d22a599
Reviewed-on: https://code.wireshark.org/review/28076
Petri-Dish: Gerald Combs <gerald@wireshark.org>
Tested-by: Petri Dish Buildbot
Reviewed-by: Gerald Combs <gerald@wireshark.org>