Do bounds checking of the offset and length in proto_tree_add_string().
Throw an exception if they don't correspond to data available in the
packet - and do so even if the protocol tree argument is null, so that
we catch very long strings that could cause the offset to overflow.

Ask why we try to handle a null pointer passed as the string argument,
while we're at it.

Bug: 14738
Change-Id: I2fa79ad0dcd1f41608844a573e045197ac60aa62
Reviewed-on: https://code.wireshark.org/review/28179
Reviewed-by: Guy Harris <guy@alum.mit.edu>
(cherry picked from commit 0fbb5f84d00426040bd37dd89d74e2ecda5b598e)
Reviewed-on: https://code.wireshark.org/review/28180