Clean up handling of counted blobs of bytes.
Don't treat the count+blob as itself a blob of bytes; use FT_NONE.
Create it with an unknown length (-1, meaning "to end of packet, for
now"), and set its length once we've finished dissecting it.  Dissect
the raw bytes of a prefixed-bytes item regardless of whether we're
building a protocol tree or not.

This means we do a better job of handling a too-large length; instead of
overflowing the offset, we throw an exception and stop dissecting, so we
don't run the risk of looping infinitely.

Bug: 14841
Change-Id: I593be9b6ba9aa15d8529f96458e53b85ace6402a
Reviewed-on: https://code.wireshark.org/review/28228
Reviewed-by: Guy Harris <guy@alum.mit.edu>
(cherry picked from commit a03eacc7aabd04a6fd0db978e0d7597220ac0515)
Reviewed-on: https://code.wireshark.org/review/28231