Do some more checks.
Don't assume that the amount of data remaining in the packet is greater
than or equal to the amount of padding; it might not be.  Throw an
exception if it isn't.

Make sure data eists before allocating a buffer for it and copying with
tvb_memcpy(), in case the length is too big.

Change-Id: Icde8b8cc687b7b9229756aabddb246f9b98c3a22
Ping-Bug: 14999
Reviewed-on: https://code.wireshark.org/review/28810
Reviewed-by: Guy Harris <guy@alum.mit.edu>
(cherry picked from commit f99d2d2afe4b92a00f3789de572a4183f226fed1)
Reviewed-on: https://code.wireshark.org/review/28811