We should be reporting expert infos on invalid radiotap headers.
Note the places where we should be doing so.

Also, note that two of the tests in ieee80211_radiotap_iterator_init()
are redundant in Wireshark, and that it's irrelevant what versions of
radiotap Linux supports - this is Wireshark code, not Linux code.  (If
there's anything to note, it's that there *is* no radiotap version other
than 0, so there's nothing *to* support.)

Change-Id: Ieabef703638b30649a097269d684f60e79db8ba5
Reviewed-on: https://code.wireshark.org/review/29031
Reviewed-by: Guy Harris <guy@alum.mit.edu>